Brute force cleanup
