--- UIAL-Main/VB/Character.cs
@@ -1,13 +1,0 @@
-﻿namespace zuoanqh.UIAL.VB
-{
-  /// <summary>
-  /// Model for character.txt files.
-  /// </summary>
-  public class CharacterTxt
-  {
-    public CharacterTxt(string fPath)
-    {
-      
-    }
-  }
-}
