+++ UIAL/VoiceBank/Character.cs
@@ -1,0 +1,12 @@
+﻿namespace zuoanqh.UIAL.VoiceBank
+{
+    /// <summary>
+    ///     Model for character.txt files.
+    /// </summary>
+    public class CharacterTxt
+    {
+        public CharacterTxt(string fPath)
+        {
+        }
+    }
+}
